feat(hooks): move rsc hook logic out of app and into packages
