+++ .gitignore
@@ -155,6 +155,7 @@ apps/api/corpus/generated/
 
 # Superpowers local workflow scratch files
 .superpowers/
+docs/superpowers/
 
 # Generated tarot card components
 # packages/cards/src/cards/
