+++ packages/hooks/index.tsx
@@ -3,8 +3,10 @@ export { trimTrailingSlashes } from './src/common/api-response';
 export * from './src/readings/reading-contracts';
 export * from './src/readings/reading-api-client';
 export * from './src/readings/reading-requests';
+export * from './src/readings/use-rsc-reading-history';
 export * from './src/avatars/avatar-contracts';
 export * from './src/avatars/avatar-api-client';
+export * from './src/avatars/use-rsc-avatar-image';
 export { default as useAvatarImage } from './src/account/use-avatar-image';
 export { default as useForgotPasswordForm } from './src/account/use-forgot-password-form';
 export { default as useInstructions } from './src/reading/use-instructions';
